Develop

+++ src/main/java/com/reviewing/review/review/domain/ReviewResponseDto.java
@@ -1,8 +1,10 @@
 package com.reviewing.review.review.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
+import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -13,14 +15,14 @@ public class ReviewResponseDto {
     private Long id;
     private String nickname;
     private String contents;
-    private float rating;
+    private BigDecimal rating;
     private int likes;
     private int dislikes;
     private boolean liked = false;
     private boolean disliked = false;
     private String createdAt;
 
-    public ReviewResponseDto(Long id, String nickname, String contents, float rating,
+    public ReviewResponseDto(Long id, String nickname, String contents, BigDecimal rating,
             int likes, int dislikes, LocalDateTime createdAt) {
         this.id = id;
         this.nickname = nickname;
